requirements.txt
aiogram>=3.20.0
speech_recognition
pydub
python-dotenv 